+++ apps/nextjs-app/.env.example
@@ -89,6 +89,9 @@ BACKEND_OIDC_OTHER={"scope": ["email", "profile"]}
 # separated by ','
 SOCIAL_AUTH_PROVIDERS=github,google,oidc
 
+# disable all endpoints related to password login (oauth & oidc work as always), default is false
+# PASSWORD_LOGIN_DISABLED=true
+
 # email configs
 BACKEND_MAIL_HOST=example@gmail.com
 BACKEND_MAIL_PORT=465
